packages/mcp/pyproject.toml
[project]
name = "fnewsteer-mcp"
version = "0.1.0"
description = "FNEWSTEER MCP server — news loss-prevention tools for AI agents"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "python-dotenv>=1.0.0",
    "fnewsteer-api",
]

[tool.uv.sources]
fnewsteer-api = { workspace = true }